+++ src/OrdersApi/Builder/Util/AmountProvider.php
@@ -75,16 +75,17 @@ class AmountProvider
         float $amountValue
     ): Breakdown {
         $accumulatedAmountValue = 0.0;
+        $accumulatedTaxValue = 0.0;
         $newItems = [];
-        $taxesCorrection = 0.0;
 
         foreach ($items as $item) {
             $itemUnitAmount = (float) $item->getUnitAmount()->getValue();
             if ($itemUnitAmount >= 0.0) {
                 $accumulatedAmountValue += $item->getQuantity() * $itemUnitAmount;
                 $newItems[] = $item;
-            } elseif ($item->hasTax()) {
-                $taxesCorrection += (float) $item->getTax()->getValue();
+                if ($item->hasTax()) {
+                    $accumulatedTaxValue += $item->getQuantity() * (float) $item->getTax()->getValue();
+                }
             }
         }
         $purchaseUnit->setItems($newItems);
@@ -100,13 +101,7 @@ class AmountProvider
 
         $taxTotal = new TaxTotal();
         $taxTotal->setCurrencyCode($currencyCode);
-        if ($isNet) {
-            $taxTotal->setValue($this->priceFormatter->formatPrice(
-                $taxes->getAmount() - ($shippingCosts->getCalculatedTaxes()->getAmount() + $taxesCorrection)
-            ));
-        } else {
-            $taxTotal->setValue($this->priceFormatter->formatPrice(0.0));
-        }
+        $taxTotal->setValue($this->priceFormatter->formatPrice($accumulatedTaxValue));
         $accumulatedAmountValue += (float) $taxTotal->getValue();
 
         $discount = new Discount();